pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dpo"
version = "0.1.0"
description = "Post-Training Alignment for Agentic AI using DPO"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.42.0",
    "accelerate>=0.29.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "sympy>=1.12",
    "scipy>=1.10.0",
    "jinja2>=3.1.0",
    "networkx>=3.0",
    "wandb>=0.16.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"
log_cli = true
pythonpath = "."

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
